[Refactor] BlockStorage/v1 -> EVS/v1 (#400)
[Refactor] BlockStorage/v1 -> EVS/v1

Move all API methods and related structures to a separate file.
Inline all Extracts and use the extract package instead.
Remove all Builders
Remove WaitForStatus / IDFromName
Move BlockStorage/v1 to EVS/v1
