--- openstack/blockstorage/v1/apiversions/requests.go
@@ -1,20 +1,0 @@
-package apiversions
-
-import (
-	"github.com/opentelekomcloud/gophertelekomcloud"
-	"github.com/opentelekomcloud/gophertelekomcloud/pagination"
-)
-
-// List lists all the Cinder API versions available to end-users.
-func List(c *golangsdk.ServiceClient) pagination.Pager {
-	return pagination.NewPager(c, listURL(c), func(r pagination.PageResult) pagination.Page {
-		return APIVersionPage{pagination.SinglePageBase(r)}
-	})
-}
-
-// Get will retrieve the volume type with the provided ID. To extract the volume
-// type from the result, call the Extract method on the GetResult.
-func Get(client *golangsdk.ServiceClient, v string) (r GetResult) {
-	_, r.Err = client.Get(getURL(client, v), &r.Body, nil)
-	return
-}
