--- openstack/blockstorage/v1/apiversions/testing/doc.go
@@ -1,2 +1,0 @@
-// apiversions_v1
-package testing
