--- openstack/blockstorage/v1/snapshots/testing/doc.go
@@ -1,2 +1,0 @@
-// snapshots_v1
-package testing
