--- openstack/blockstorage/v1/snapshots/util.go
@@ -1,22 +1,0 @@
-package snapshots
-
-import (
-	"github.com/opentelekomcloud/gophertelekomcloud"
-)
-
-// WaitForStatus will continually poll the resource, checking for a particular
-// status. It will do this for the amount of seconds defined.
-func WaitForStatus(c *golangsdk.ServiceClient, id, status string, secs int) error {
-	return golangsdk.WaitFor(secs, func() (bool, error) {
-		current, err := Get(c, id).Extract()
-		if err != nil {
-			return false, err
-		}
-
-		if current.Status == status {
-			return true, nil
-		}
-
-		return false, nil
-	})
-}
