--- openstack/blockstorage/v1/volumes/doc.go
@@ -1,5 +1,0 @@
-// Package volumes provides information and interaction with volumes in the
-// OpenStack Block Storage service. A volume is a detachable block storage
-// device, akin to a USB hard drive. It can only be attached to one instance at
-// a time.
-package volumes
