--- openstack/blockstorage/v1/volumetypes/urls.go
@@ -1,19 +1,0 @@
-package volumetypes
-
-import "github.com/opentelekomcloud/gophertelekomcloud"
-
-func listURL(c *golangsdk.ServiceClient) string {
-	return c.ServiceURL("types")
-}
-
-func createURL(c *golangsdk.ServiceClient) string {
-	return listURL(c)
-}
-
-func getURL(c *golangsdk.ServiceClient, id string) string {
-	return c.ServiceURL("types", id)
-}
-
-func deleteURL(c *golangsdk.ServiceClient, id string) string {
-	return getURL(c, id)
-}
